Create Reply proxies at the time of use
With main and reply interfaces merged, there is no longer an R suffix
to differentiate calls from replies. While App already disallows dead
Msger resurrection by non-creators, the dead mrid can get reassigned,
and the Reply message could be delivered there. With this commit the
creator id is reset to msger_id() in on_msger_destroyed, and if the
reply proxy is then created at the time of use from creator_link(),
it is bounced back to the sender.