Unify socket path construction
Replace code for creating user-local or system-local sockets with a
unified socket_path_from_name and the ~socketname prefix to indicate
user-local sockets. This allows path construction from just the
information already in the interface, and the reuse of this code on
both the connect and the bind side. A feature of expanding environment
variables in the socket name has also been added, allowing separation
of session socket instances.