Implement automatic import connections
App will now create Extern connections for imported interface messages.
The messages are queued at the new Extern object until connection is
established, at which point they are returned to App for reprocessing.
The queueing ensures only one connection is pending at a time, to avoid
launching multiple server processes.

An additional security check was added to server processes listening on
abstract sockets, restricting incoming connections to processes with a
specific uid. The filter uid is obtained by checking the ownership of
the real path of the socket. That is usually $XDG_RUNTIME_DIR, owned
by the user, so the incoming connections are restricted to that uid.